requirements.txt
# DEPLOYMENT
django == 3.1.4
python-decouple == 3.3          # needed to parse secrets from .env
django-crispy-forms~=1.10.0

# TESTING
coverage == 5.3
